Fixed export error in module 1

+++ pages/1_Module_1_Generate_Image_Mosaic.py
@@ -350,7 +350,7 @@ if st.button("Search Landsat Imagery", type="primary") and st.session_state.aoi 
         #Create and image composite/mosaic for thermal bands
         thermal_median = thermal_collection.median().clip(aoi)
         #composite for multispectral data
-        composite = collection.median().clip(aoi).addBands(thermal_median)
+        composite = collection.median().clip(aoi).addBands(thermal_median).toFloat()
         # Store in session state for use in other modules
         st.session_state['composite'] = composite
         st.session_state['Image_metadata'] = detailed_stats
